done: the authentication is done using spring security and jwt mechanism

+++ src/main/java/dev/codex/redindiansnight/Blog/Application/Dtos/Requests/BlogRequest.java
@@ -17,6 +17,6 @@ public record BlogRequest(
         String content,
 
         @NotNull
-        Long userId
+        Integer userId
 ) implements Serializable {
 }
